feat: Don't try to follow seek to the beginning
Sometimes, when opening a new tab to add a new video,
that fires a 'seeked' event, which pulls me back from lap whichever
to the beginning of the session. Sad times.

This might not be the right solution. You might want to start all of the
videos together at the beginning. I might need to track the age of the
page instead.